Creacion de una nueva rama y modificacion de esta para que quede exactamente igual que en la pull request
