+++ src/com/cjimenezro/Electricidad_Chispas/Servicios.java
@@ -1,4 +1,4 @@
-package com.cjimenezro.Electricidad_Chispas.domain.models;
+package com.cjimenezro.Electricidad_Chispas;
 
 public class Servicios implements ProductosServicios{
     private int codigo,precio,tipoIva;
