--- src/com/cjimenezro/Electricidad_Chispas/domain/usecase/UpdateCustomerUseCase.java
@@ -1,17 +1,0 @@
-package com.cjimenezro.Electricidad_Chispas.domain.usecase;
-
-import com.cjimenezro.Electricidad_Chispas.data.CustomerDataStore;
-import com.cjimenezro.Electricidad_Chispas.domain.models.Clientes;
-
-public class UpdateCustomerUseCase {
-
-    private CustomerDataStore customerDataStore;
-
-    public UpdateCustomerUseCase(CustomerDataStore customerDataStore){
-        this.customerDataStore = customerDataStore;
-    }
-
-    public void execute(Clientes customer){
-        customerDataStore.updateCustomer(customer);
-    }
-}
